+++ examples/sync/get_timings.py
@@ -1,13 +1,13 @@
 from aladhan import (
     Client,
-    TimingsDateArg,
-    Parameters,
-    methods,
-    Tune,
-    Schools,
-    MidnightModes,
     LatitudeAdjustmentMethods,
+    MidnightModes,
+    Parameters,
+    Schools,
     Timings,
+    TimingsDateArg,
+    Tune,
+    methods,
 )